refactor: requests 관심사 별 분리 & query 훅 분리
