--- src/api/requests.ts
@@ -1,70 +1,0 @@
-import type {
-  UserInfoData,
-  UserInfoProps,
-  ProductInfo,
-  RankingApiProps,
-  OrderInfoValues,
-  ProductSummaryData,
-  ThemeInfo,
-  ThemeIdInfoData,
-  ThemeIdItemsData,
-  OrderResponseData,
-  ProductDetailData,
-  ProductWishData,
-  ProductReviewData,
-} from '@/types';
-import { apiClient } from './apiClient';
-
-interface FetchOrderProps {
-  orderData: OrderInfoValues;
-  index: number;
-}
-interface fetchThemeIdItemsProps {
-  index: number;
-  currentCursor: number;
-}
-
-export const requests = {
-  fetchUserInfos: ({ username, password }: UserInfoProps): Promise<UserInfoData> => {
-    const data = {
-      email: username.value,
-      password: password.value,
-    };
-    return apiClient.post('/api/login', data);
-  },
-  fetchOrder: ({ orderData, index }: FetchOrderProps): Promise<OrderResponseData> => {
-    const { message, name, receiverInfos } = orderData;
-    const data = {
-      productId: index,
-      message: message,
-      messageCardId: 'card123',
-      ordererName: name,
-      receivers: receiverInfos,
-    };
-    return apiClient.post('/api/order', data);
-  },
-  fetchSummary: (index: number): Promise<ProductSummaryData> => {
-    return apiClient.get(`/api/products/${index}/summary`);
-  },
-  fetchRanking: ({ activeGeneration, activeFilter }: RankingApiProps): Promise<ProductInfo[]> => {
-    return apiClient.get(
-      `/api/products/ranking?targetType=${activeGeneration}&rankType=${activeFilter}`
-    );
-  },
-  fetchTheme: (): Promise<ThemeInfo[]> => apiClient.get('/api/themes'),
-  fetchThemeIdInfo: (index: number): Promise<ThemeIdInfoData> =>
-    apiClient.get(`/api/themes/${index}/info`),
-  fetchThemeIdItems: ({
-    index,
-    currentCursor,
-  }: fetchThemeIdItemsProps): Promise<ThemeIdItemsData> =>
-    apiClient.get(`/api/themes/${index}/products?cursor=${currentCursor}`),
-
-  fetchProduct: (index: number): Promise<ProductInfo> => apiClient.get(`/api/products/${index}`),
-  fetchProductDetail: (index: number): Promise<ProductDetailData> =>
-    apiClient.get(`/api/products/${index}/detail`),
-  fetchProductReview: (index: number): Promise<ProductReviewData> =>
-    apiClient.get(`/api/products/${index}/highlight-review`),
-  fetchProductWish: (index: number): Promise<ProductWishData> =>
-    apiClient.get(`/api/products/${index}/wish`),
-};
