+++ src/page/Home/hooks/useRanking.ts
@@ -1,11 +1,11 @@
-import { requests } from '@/api/requests';
+import { productRequests } from '@/api/productRequests';
 import type { ProductInfo, RankingApiProps } from '@/types';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 const useRanking = ({ activeGeneration, activeFilter }: RankingApiProps) => {
   const { data } = useSuspenseQuery<ProductInfo[]>({
     queryKey: ['rankDatas', activeGeneration, activeFilter],
-    queryFn: () => requests.fetchRanking({ activeGeneration, activeFilter }),
+    queryFn: () => productRequests.fetchRanking({ activeGeneration, activeFilter }),
   });
 
   return data;
