+++ src/page/Home/hooks/useThemes.ts
@@ -1,11 +1,11 @@
-import { requests } from '@/api/requests';
+import { themeRequests } from '@/api/themeRequests';
 import type { ThemeInfo } from '@/types';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 const useThemes = () => {
   const { data } = useSuspenseQuery<ThemeInfo[]>({
     queryKey: ['themeDatas'],
-    queryFn: requests.fetchTheme,
+    queryFn: themeRequests.fetchTheme,
   });
 
   return data;
