+++ src/page/Product/components/ProductReviewSection.tsx
@@ -1,20 +1,16 @@
-import { requests } from '@/api/requests';
-import { useSuspenseQuery } from '@tanstack/react-query';
+import useProductReviewQuery from '../hooks/useProductReviewQuery';
 
 interface ProductReviewSectionProps {
   index: number;
 }
 const ProductReviewSection = ({ index }: ProductReviewSectionProps) => {
-  const productReviewQuery = useSuspenseQuery({
-    queryKey: ['productReviewData', index],
-    queryFn: () => requests.fetchProductReview(index),
-  });
+  const datas = useProductReviewQuery(index);
 
   return (
     <div>
       <div>
         <div>
-          {productReviewQuery.data?.reviews.map(review => (
+          {datas.reviews.map(review => (
             <div key={review.id}>
               <p>{review.authorName}</p>
               <p>{review.content}</p>
