+++ src/page/Product/hooks/useProductReviewQuery.ts
@@ -1,0 +1,12 @@
+import { productRequests } from '@/api/productRequests';
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+const useProductReviewQuery = (index: number) => {
+  const { data } = useSuspenseQuery({
+    queryKey: ['productReviewData', index],
+    queryFn: () => productRequests.fetchProductReview(index),
+  });
+  return data;
+};
+
+export default useProductReviewQuery;
