+++ src/page/Themes/hooks/useInfiniteApiQuery.ts
@@ -1,6 +1,6 @@
-import { useInfiniteQuery } from "@tanstack/react-query";
-import { useParamsIndex } from "../../../hooks/useParamsIndex";
-import { requests } from "@/api/requests";
+import { useInfiniteQuery } from '@tanstack/react-query';
+import { useParamsIndex } from '../../../hooks/useParamsIndex';
+import { themeRequests } from '@/api/themeRequests';
 
 export const useInfiniteApiQuery = () => {
   const index = useParamsIndex();
@@ -8,7 +8,8 @@ export const useInfiniteApiQuery = () => {
   const themeItemsQuery = useInfiniteQuery({
     queryKey: ['themeIdItems', index],
     initialPageParam: 0,
-    queryFn: ({ pageParam = 0 }) => requests.fetchThemeIdItems({ index, currentCursor: pageParam }),
+    queryFn: ({ pageParam = 0 }) =>
+      themeRequests.fetchThemeIdItems({ index, currentCursor: pageParam }),
     getNextPageParam: lastPage => (lastPage.hasMoreList ? lastPage.cursor : undefined),
   });
   const items = themeItemsQuery.data?.pages.flatMap(page => page.list) || [];
