+++ src/page/Themes/hooks/useThemeInfoQuery.ts
@@ -1,11 +1,11 @@
-import { requests } from '@/api/requests';
+import { themeRequests } from '@/api/themeRequests';
 import type { ThemeIdInfoData } from '@/types';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 const useThemeInfoQuery = (index: number) => {
   return useSuspenseQuery<ThemeIdInfoData>({
     queryKey: ['themeIdInfoData', index],
-    queryFn: () => requests.fetchThemeIdInfo(index),
+    queryFn: () => themeRequests.fetchThemeIdInfo(index),
   });
 };
